refactor(cloud): centralize anonymous spend gates
